refactor: 컴포넌트 폴더 구조 개선
- 컴포넌트를 도메인별/역할별로 분류
- 앱 설정 파일을 app/ 폴더로 분리
- import 경로 절대 경로로 변경
- vite.config.js에 path alias 설정 추가
