--- frontend/src/App.tsx
@@ -1,13 +1,0 @@
-import { Routes, Route } from "react-router-dom";
-import DefaultLayout from "./layouts/DefaultLayout";
-import Home from "./pages/Home";
-
-export default function App() {
-  return (
-    <Routes>
-      <Route path="/" element={<DefaultLayout />}>
-        <Route index element={<Home />} />
-      </Route>
-    </Routes>
-  );
-}
