+++ frontend/src/components/accommodation/AccommodationCard.jsx
