+++ frontend/src/components/accommodation/AmenitiyModal.jsx
