+++ frontend/src/components/accommodation/HotAccommodation.jsx
