+++ frontend/src/components/accommodation/HotPoint.jsx
