+++ frontend/src/components/accommodation/ResultHeader.jsx
