+++ frontend/src/hooks/useFilteredAccommodations.js
@@ -1,8 +1,8 @@
 import { useMemo } from "react";
 import { useLocation } from "react-router-dom";
-import { ACCOMMODATIONS } from "../data/accommodations";
-import { CATEGORY_OPTIONS, PRICE_MIN, PRICE_MAX } from "../constants/filters";
-import { filterAccommodationsByKeyword } from "../utils/searchUtils";
+import { ACCOMMODATIONS } from "@data/accommodations";
+import { CATEGORY_OPTIONS, PRICE_MIN, PRICE_MAX } from "@constants/filters";
+import { filterAccommodationsByKeyword } from "@utils/searchUtils";
 
 export default function useFilteredAccommodations(type, keyword, keywordSlug) {
   const location = useLocation();
